packages/scraper/pyproject.toml
[project]
name = "scraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Matthias", email = "matthias@repath.earth" }
]
requires-python = ">=3.13"
dependencies = [
    "awswrangler>=3.11.0",
    "country-converter>=1.3",
    "duckdb>=1.2.2",
    "mlb-statsapi>=1.9.0",
    "pandas>=2.2.3",
    "prefect>=3.4.1",
    "prefect-cloud>=0.1.6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
